Services y Controller #51

+++ backend/buenGusto-final/src/main/java/com/utn/app/buenGusto/articuloInsumo/ArticuloInsumoRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ArticuloInsumoRepository extends JpaRepository<ArticuloInsumoEntity, Integer>{}
+public interface ArticuloInsumoRepository extends JpaRepository<ArticuloInsumoEntity, Long>{}
