+++ backend/buenGusto-final/src/main/java/com/utn/app/buenGusto/articuloManufacturado/ArticuloManufacturadoRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ArticuloManufacturadoRepository extends JpaRepository<ArticuloManufacturadoEntity, Integer> {
+public interface ArticuloManufacturadoRepository extends JpaRepository<ArticuloManufacturadoEntity, Long> {
 
 	/*
 	 * @Query("SELECT * FROM articulo art,articulo_manufacturado artM WHERE artM.articulo_id = art.id"
