+++ backend/buenGusto-final/src/main/java/com/utn/app/buenGusto/domicilio/DomicilioRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DomicilioRepository extends JpaRepository<DomicilioEntity, Integer> {
+public interface DomicilioRepository extends JpaRepository<DomicilioEntity, Long> {
 }
