+++ backend/buenGusto-final/src/main/java/com/utn/app/buenGusto/historicoEstado/HistoricoEstadoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HistoricoEstadoRepository extends JpaRepository<HistoricoEstadoEntity, Integer>{
+public interface HistoricoEstadoRepository extends JpaRepository<HistoricoEstadoEntity, Long>{
 
 }
