+++ backend/buenGusto-final/src/main/java/com/utn/app/buenGusto/pedido/PedidoRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PedidoRepository extends JpaRepository<PedidoEntity, Integer>{}
+public interface PedidoRepository extends JpaRepository<PedidoEntity, Long>{}
