+++ backend/buenGusto-final/src/main/java/com/utn/app/buenGusto/stockArticulo/StockArticuloRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StockArticuloRepository extends JpaRepository<StockArticuloEntity, Integer>{}
+public interface StockArticuloRepository extends JpaRepository<StockArticuloEntity, Long>{}
