+++ backend/buenGusto-final/src/main/java/com/utn/app/buenGusto/usuario/UsuarioRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<UsuarioEntity, Integer>{}
+public interface UsuarioRepository extends JpaRepository<UsuarioEntity, Long>{}
